test(tools): file the embedding tests under the module they drive
The test-file naming gate wants every _test.go named after a module it
tests, and the embedding tests had a name of their own. They drive the
individual surface's dispatcher first, so they live in
individual_catalog_test.go now, with the meta and dynamic surfaces held
to the same answer from there.
